+++ test/flutter_dev_utils_test.dart
@@ -1,10 +1,14 @@
+import 'package:flutter_dev_utils/flutter_dev_utils.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  test('test_async_try_catch_handler', () {
-    expect('a', 'a');
+  test('test_async_try_handler', () async {
+    final result = await asyncTryHandler(tryFunction: () async {
+      return await Future.delayed(const Duration(milliseconds: 50), () => 'a');
+    });
+    expectLater('a', result);
   });
-  test('test_sync_try_catch_handler', () {
+  test('test_sync_try_handler', () {
     expect('a', 'a');
   });
 }